Move building properties to own folder
