+++ cea/demand/building_properties/__init__.py
@@ -1,0 +1,1 @@
+from .building_properties import BuildingProperties
