+++ cea/demand/building_properties/building_geometry.py
