+++ cea/demand/building_properties/building_solar.py
