+++ cea/demand/building_properties/building_typology.py
